requirements.txt
Django
python-dotenv
djangorestframework
g4f
